Rename SlidedownHtml.ts to SlidedownElement.ts
After shifting from hard-coded HTML approach to Web API approach, the function returns a web Element so we are renaming it here to be more clear.
